Rename package from `bigdbm` to `real_intent`.
