--- bigdbm/analyze/__init__.py
@@ -1,2 +1,0 @@
-"""Generate string insights based on lead data."""
-from bigdbm.analyze.insights import OpenAIInsightsGenerator, ValidatedInsightsGenerator
