--- bigdbm/deliver/followupboss/__init__.py
@@ -1,3 +1,0 @@
-"""Deliverers to interact with Follow Up Boss."""
-from bigdbm.deliver.followupboss.vanilla import FollowUpBossDeliverer
-from bigdbm.deliver.followupboss.ai_fields import AIFollowUpBossDeliverer
