--- bigdbm/process/__init__.py
@@ -1,3 +1,0 @@
-"""Series of processors to handle the processing flow in different ways."""
-from bigdbm.process.simple import SimpleProcessor
-from bigdbm.process.fill import FillProcessor
