+++ real_intent/analyze/__init__.py
@@ -1,0 +1,2 @@
+"""Generate string insights based on lead data."""
+from real_intent.analyze.insights import OpenAIInsightsGenerator, ValidatedInsightsGenerator
