+++ real_intent/analyze/base.py
