+++ real_intent/deliver/__init__.py
@@ -1,2 +1,2 @@
 """Take a list of MD5WithPII (leads) and format/deliver them. Do something with them, essentially."""
-from bigdbm.deliver.csv import CSVStringFormatter
+from real_intent.deliver.csv import CSVStringFormatter
