+++ real_intent/deliver/followupboss/ai_fields.py
@@ -5,10 +5,10 @@ from pydantic import BaseModel, Field
 import json
 from typing import Literal, Any
 
-from bigdbm.schemas import MD5WithPII
-from bigdbm.deliver.followupboss.vanilla import FollowUpBossDeliverer, EventType
-from bigdbm.deliver.followupboss.ai_prompt import SYSTEM_PROMPT
-from bigdbm.internal_logging import log, log_span
+from real_intent.schemas import MD5WithPII
+from real_intent.deliver.followupboss.vanilla import FollowUpBossDeliverer, EventType
+from real_intent.deliver.followupboss.ai_prompt import SYSTEM_PROMPT
+from real_intent.internal_logging import log, log_span
 
 
 class CustomField(BaseModel):
