+++ real_intent/process/__init__.py
@@ -1,0 +1,3 @@
+"""Series of processors to handle the processing flow in different ways."""
+from real_intent.process.simple import SimpleProcessor
+from real_intent.process.fill import FillProcessor
