+++ real_intent/validate/base.py
@@ -1,8 +1,8 @@
 """Base abstract interface for validators."""
 from abc import ABC, abstractmethod
 
-from bigdbm.schemas import MD5WithPII
-from bigdbm.internal_logging import log, log_span
+from real_intent.schemas import MD5WithPII
+from real_intent.internal_logging import log, log_span
 
 
 class BaseValidator(ABC):
