+++ real_intent/validate/home_attrs.py
@@ -1,6 +1,6 @@
 """Validations for attributes related to home ownership or households."""
-from bigdbm.schemas import MD5WithPII
-from bigdbm.validate.base import BaseValidator
+from real_intent.schemas import MD5WithPII
+from real_intent.validate.base import BaseValidator
 
 
 class NotRenterValidator(BaseValidator):
